+++ nokhwa-core/src/types.rs
@@ -947,7 +947,7 @@ impl KnownCameraControl {
 
 impl Display for KnownCameraControl {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{:?}", &self)
+        write!(f, "{self:?}")
     }
 }